requirements.txt
discord.py==2.1.1
python-dotenv==0.21.1
google-api-python-client==2.78.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
gspread==5.7.2
matplotlib==3.7.0
apscheduler==3.10.0
requests==2.28.2
tcp-latency
numpy<2
dateparser
unidecode